requirements.txt
numpy==1.24.4
pandas==2.0.3
darts==0.24.0
streamlit==1.26.0
tk==0.1.0
stqdm==0.0.5